M7 RAG_Retrieval Augmented Generation/Context Management & Langchain & API/Langchain_Chat with Memory/Langchain_ Chat/requirements.txt
# LangChain Chat with Memory - Production Dependencies
# Last updated: 2026-03-31

# Core LangChain framework
langchain>=0.2,<0.3
langchain-openai>=0.1.5,<0.2
langchain-core>=0.2,<0.3

# API & Web
fastapi>=0.110,<0.120
uvicorn[standard]>=0.27,<0.28
gradio>=4.40,<5.0

# LLM & External APIs
openai>=1.10,<2.0

# Utilities
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
pydantic>=2.0,<3.0

# Optional: for better performance
# gunicorn>=21.0  # For production ASGI serving
# python-multipart>=0.0.6  # For form data handling
# slowapi>=0.1.9  # For rate limiting